move some more tests to shared
