+++ composite/integration-tests/src/testShared/kotlin/com/apollographql/apollo3/ResponseNormalizationTest.kt
